* Rewire RecordPattern to extend Pattern rather than TypePattern
* Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/2043
