+++ org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/ast/Pattern.java
@@ -17,6 +17,7 @@ import org.eclipse.jdt.internal.compiler.codegen.BranchLabel;
 import org.eclipse.jdt.internal.compiler.codegen.CodeStream;
 import org.eclipse.jdt.internal.compiler.lookup.BlockScope;
 import org.eclipse.jdt.internal.compiler.lookup.MethodBinding;
+import org.eclipse.jdt.internal.compiler.lookup.Scope;
 import org.eclipse.jdt.internal.compiler.lookup.TypeBinding;
 
 public abstract class Pattern extends Expression {
@@ -52,7 +53,9 @@ public abstract class Pattern extends Expression {
 	 * @return whether pattern covers the given type or not
 	 */
 	public boolean coversType(TypeBinding type) {
-		return false;
+		if (type == null || this.resolvedType == null)
+			return false;
+		return (type.isSubtypeOf(this.resolvedType, false));
 	}
 
 	public boolean isAlwaysTrue() {
@@ -61,12 +64,39 @@ public abstract class Pattern extends Expression {
 
 	public abstract void generateCode(BlockScope currentScope, CodeStream codeStream, BranchLabel trueLabel, BranchLabel falseLabel);
 
+	@Override
+	public boolean checkUnsafeCast(Scope scope, TypeBinding castType, TypeBinding expressionType, TypeBinding match, boolean isNarrowing) {
+		if (!castType.isReifiable())
+			return CastExpression.checkUnsafeCast(this, scope, castType, expressionType, match, isNarrowing);
+		else
+			return super.checkUnsafeCast(scope, castType, expressionType, match, isNarrowing);
+	}
+
 	public TypeReference getType() {
 		return null;
 	}
 
 	// 14.30.3 Properties of Patterns: A pattern p is said to be applicable at a type T if ...
-	protected abstract boolean isApplicable(TypeBinding other, BlockScope scope);
+	protected boolean isApplicable(TypeBinding other, BlockScope scope) {
+		TypeBinding patternType = this.resolvedType;
+		if (patternType == null) // ill resolved pattern
+			return false;
+		// 14.30.3 Properties of Patterns doesn't allow boxing nor unboxing, primitive widening/narrowing.
+		if (patternType.isBaseType() != other.isBaseType()) {
+			scope.problemReporter().incompatiblePatternType(this, other, patternType);
+			return false;
+		}
+		if (patternType.isBaseType()) {
+			if (!TypeBinding.equalsEquals(other, patternType)) {
+				scope.problemReporter().incompatiblePatternType(this, other, patternType);
+				return false;
+			}
+		} else if (!checkCastTypesCompatibility(scope, other, patternType, null, true)) {
+			scope.problemReporter().incompatiblePatternType(this, other, patternType);
+			return false;
+		}
+		return true;
+	}
 
 	public abstract boolean dominates(Pattern p);
